+++ game_patch/graphics/d3d11/gr_d3d11_context.h
@@ -10,6 +10,7 @@
 #include "gr_d3d11_state.h"
 #include "../../misc/alpine_settings.h"
 #include "../../rf/gr/gr.h"
+#include "gr_d3d11_mesh.h"
 
 namespace df::gr::d3d11
 {
@@ -87,7 +88,7 @@ namespace df::gr::d3d11
             bool fog_allowed = mode.get_fog_type() != gr::FOG_NOT_ALLOWED;
             int colorblind_mode = g_alpine_game_config.colorblind_mode;
             float dynamic_light_ndotl = g_alpine_game_config.dynamic_light_ndotl;
-            float pixel_light_overbright = g_alpine_game_config.pixel_light_overbright;
+            float pixel_light_overbright = g_level_pixel_light_overbright;
             if (force_update_ || current_alpha_test_ != alpha_test || current_fog_allowed_ != fog_allowed || current_color_ != color || current_colorblind_mode_ != colorblind_mode || current_lightmap_only_ != lightmap_only || current_dynamic_lighting_ != dynamic_lighting || current_self_illumination_ != self_illumination || current_apply_light_scale_ != apply_light_scale || current_dynamic_light_ndotl_ != dynamic_light_ndotl || current_pixel_light_overbright_ != pixel_light_overbright) {
                 current_alpha_test_ = alpha_test;
                 current_fog_allowed_ = fog_allowed;